update figure to version 4
